+++ packages/uiengine-theme-default/src/vue/components/ContentLabel.vue
@@ -23,7 +23,7 @@ export default {
   padding var(--space-xxs) var(--space-s)
   border 1px solid var(--color-contentheader-label-bg)
   border-radius var(--space-m)
-  border-bottom-right-radius 0
+  border-bottom-left-radius 0
   color var(--color-contentheader-label-text)
   background var(--color-contentheader-label-bg)
   font-size var(--font-size-s)