+++ packages/uiengine-theme-default/src/vue/components/ContentTag.vue
@@ -1,0 +1,38 @@
+<template>
+  <router-link
+    class="tag"
+    :to="target"
+  >{{ tag }}</router-link>
+</template>
+
+<script>
+export default {
+  props: {
+    tag: {
+      type: String,
+      required: true
+    }
+  },
+
+  computed: {
+    target () {
+      return {
+        name: 'search',
+        params: { query: this.tag }
+      }
+    }
+  }
+}
+</script>
+
+<style lang="stylus" scoped>
+.tag
+  position relative
+  padding var(--space-xxs) var(--space-s)
+  border 1px solid var(--color-contentheader-tag-text)
+  border-radius var(--space-m)
+  border-bottom-left-radius 0
+  color var(--color-contentheader-tag-text)
+  font-size var(--font-size-s)
+  text-decoration none
+</style>