+++ test/project/uiengine.config.js
@@ -48,7 +48,7 @@ module.exports = {
   },
 
   // Here you can configure the template that the variant preview gets embeded in.
-  variantTemplate: 'variant-preview.pug',
+  template: 'uiengine.pug',
 
   // UIengine will `require` the theme, which has to provide a defined set of functions
   // for setup and rendering. For details see the theme docs.